% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ndvi_array.R
\name{build_ndvi_array}
\alias{build_ndvi_array}
\title{build_ndvi_array}
\usage{
build_ndvi_array(files_desc)
}
\arguments{
\item{files_desc}{object returned from *build_ndvi_desc*}
}
\value{

}
\description{
build_ndvi_array
}
