% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ndvi_desc.R
\name{build_ndvi_desc}
\alias{build_ndvi_desc}
\title{build_ndvi_desc}
\usage{
build_ndvi_desc(
  folder,
  date.loc = c(15, 21),
  date.format = "\%Y\%j",
  start.date = NULL,
  end.date = NULL
)
}
\arguments{
\item{folder}{tbc}

\item{date.loc}{tbc}

\item{date.format}{tbc}

\item{start.date}{tbc}

\item{end.date}{tbc}
}
\value{

}
\description{
build_ndvi_desc
}
