% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_rain_array.R
\name{build_rain_array}
\alias{build_rain_array}
\title{Generate array of CHIRPS rainfall data}
\usage{
build_rain_array(rain_folder, rain_files)
}
\arguments{
\item{rain_folder}{the folder path where the daily chirps data are stored}

\item{rain_files}{list object of the files in that folder, can be creates using list.files()}
}
\value{

}
\description{
Generate array of CHIRPS rainfall data
}
