% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_array.R
\name{extract_array}
\alias{extract_array}
\title{extract_array}
\usage{
extract_array(array, extent, coords, third_dim_ind, resolution)
}
\arguments{
\item{array}{an array with dimensions 1 and 2 being the x and y values of the rasters layers identified in files_desc,
the 3rd dimension is multiple dates for same spatial extent}

\item{extent}{an extent object. contains xmin, xmax, ymin, ymax of layers in array}

\item{coords}{a 2 column matrix/dataframe with the x and y coordinates in same projection as extent}

\item{third_dim_ind}{index of which layer to be used, i.e. third dimension}

\item{resolution}{a vector of 2 numers indicating the cell resolution of the spatial object}
}
\value{

}
\description{
extract_array
}
