% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matrix.R
\name{extract_matrix}
\alias{extract_matrix}
\title{extract_matrix}
\usage{
extract_matrix(matrix, coords, raster_extent, raster_res)
}
\arguments{
\item{matrix}{tbc}

\item{coords}{tbc}

\item{raster_extent}{tbc}

\item{raster_res}{tbc}
}
\value{

}
\description{
extract_matrix
}
