% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi_extraction.R
\name{ndvi_extraction}
\alias{ndvi_extraction}
\title{ndvi_extraction}
\usage{
ndvi_extraction(
  files_desc,
  array,
  extent,
  resolution,
  avg_matrix = NULL,
  date,
  coords,
  cpus = 1
)
}
\arguments{
\item{files_desc}{object returned from *build_ndvi_desc*}

\item{array}{an array with dimensions 1 and 2 being the y and x values (respectively) of the rasters layers
identified in files_desc, the 3rd dimension is multiple dates for same spatial extent.
Build using *14_NDVIBuildArray*}

\item{extent}{an extent object. contains xmin, xmax, ymin, ymax of the raster layers making up the array}

\item{resolution}{a vector with 2 values indicating the x and y resolution of each raster cell}

\item{avg_matrix}{(default: NULL) if calculating anomally NDVI this is a matrix with the long-term mean NDVI values}

\item{date}{a vector of the dates}

\item{coords}{a 2 column matrix/dataframe with the x and y coordinates in same projection as extent}

\item{cpus}{(default: 1) how many cpus you have available to parallelize across}
}
\value{

}
\description{
This function matches NDVI values to spatial point data. It will return:
}
\details{
- **iNDVI** -- interpolated NDVI, giving the estimated value of NDVI if the
  date of the spatial data falls in between the dates of NDVI layers
  (which are produced at 16 day intervals)
  - **dNDVI** -- delta NDVI describes the change in NDVI between the closest
  date layer and the layer previous
  - **aNDVI** -- anomally NDVI is *(iNDVI-mNDVI)/mNDVI* where mNDVI is the long-term
  mean NDVI value for the given cell

The function is most accurate using projected data, i.e. both the NDVI files and point data should
be in UTM (or similar). The function will still work if the data are in latlong, however there may be
some cases where the values returned are associated with a neighbouring NDVI cell. This is due to
rounding issues.
}
