% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_extract_chirps.R
\name{rain_extract_chirps}
\alias{rain_extract_chirps}
\title{Extract daily rainfall from CHIRPS data}
\usage{
rain_extract_chirps(
  rain_array,
  rain_dates,
  extent,
  resolution,
  cumrain = 0,
  coords_dates,
  coords
)
}
\arguments{
\item{rain_array}{an array with dimensions 1 and 2 being the y and x values (respectively) of the rain raster/tiff layers, 
and the 3rd dimension is the multiple dates for same spatial extent.
 Build using \emph{build_rain_array}}

\item{rain_dates}{a vector of class Date containing the dates corresponding to each "layer" in the 3rd dimension of the array}

\item{extent}{an extent object. contains xmin, xmax, ymin, ymax of the raster/tiff layers used to make the array}

\item{resolution}{a vector with 2 values indicating the x and y resolution of each raster cell}

\item{cumrain}{(defualt = 0) the number of days used to calculate the cumulative rainfall at a location}

\item{coords_dates}{a vector of class Date of the dates for each of the coordinates in coords}

\item{coords}{a 2 column matrix/dataframe with the x and y coordinates in same projection as extent}
}
\value{

}
\description{
Extract daily rainfall from CHIRPS data
}
