% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matrix.R
\name{extract_matrix}
\alias{extract_matrix}
\title{extract_matrix}
\usage{
extract_matrix(matrix, coords, raster_extent, raster_res)
}
\arguments{
\item{matrix}{a matrix created from a Raster layer}

\item{coords}{a 2 column matrix/dataframe with the x and y coordinates in same projection as the Raster layer}

\item{raster_extent}{an extent object. contains xmin, xmax, ymin, ymax of the Raster layer the matrix was created from}

\item{raster_res}{a vector with 2 values indicating the x and y resolution of each Raster cell}
}
\value{

}
\description{
extract_matrix
}
